Little bit of clean up
