+++ src/partials/Program.Menus.Questions.cs
@@ -1,0 +1,45 @@
+using Pannella.Helpers;
+
+namespace Pannella;
+
+internal partial class Program
+{
+    private static void AskAboutNewCores(bool force = false)
+    {
+        while (ServiceHelper.SettingsService.GetConfig().download_new_cores == null || force)
+        {
+            force = false;
+
+            Console.WriteLine("Would you like to, by default, install new cores? [Y]es, [N]o, [A]sk for each:");
+
+            ConsoleKey response = Console.ReadKey(true).Key;
+
+            ServiceHelper.SettingsService.GetConfig().download_new_cores = response switch
+            {
+                ConsoleKey.Y => "yes",
+                ConsoleKey.N => "no",
+                ConsoleKey.A => "ask",
+                _ => null
+            };
+        }
+    }
+
+    private static bool AskYesNoQuestion(string question)
+    {
+        Console.WriteLine($"{question} [Y]es, [N]o");
+
+        bool? result = null;
+
+        while (result == null)
+        {
+            result = Console.ReadKey(true).Key switch
+            {
+                ConsoleKey.Y => true,
+                ConsoleKey.N => false,
+                _ => null
+            };
+        }
+
+        return result.Value;
+    }
+}
