+++ src/partials/Program.Menus.cs
@@ -1,10 +1,6 @@
-using System.ComponentModel;
 using ConsoleTools;
 using Pannella.Helpers;
-using Pannella.Models.DisplayModes;
 using Pannella.Models.Extras;
-using Pannella.Models.OpenFPGA_Cores_Inventory;
-using Pannella.Models.Settings;
 using Pannella.Services;
 
 namespace Pannella;
@@ -362,408 +358,6 @@ internal partial class Program
             .Add("Exit", ConsoleMenu.Close);
 
         menu.Show();
-    }
-
-    private static void AskAboutNewCores(bool force = false)
-    {
-        while (ServiceHelper.SettingsService.GetConfig().download_new_cores == null || force)
-        {
-            force = false;
-
-            Console.WriteLine("Would you like to, by default, install new cores? [Y]es, [N]o, [A]sk for each:");
-
-            ConsoleKey response = Console.ReadKey(true).Key;
-
-            ServiceHelper.SettingsService.GetConfig().download_new_cores = response switch
-            {
-                ConsoleKey.Y => "yes",
-                ConsoleKey.N => "no",
-                ConsoleKey.A => "ask",
-                _ => null
-            };
-        }
-    }
-
-    private static bool AskYesNoQuestion(string question)
-    {
-        Console.WriteLine($"{question} [Y]es, [N]o");
-
-        bool? result = null;
-
-        while (result == null)
-        {
-            result = Console.ReadKey(true).Key switch
-            {
-                ConsoleKey.Y => true,
-                ConsoleKey.N => false,
-                _ => null
-            };
-        }
-
-        return result.Value;
-    }
-
-    private static Dictionary<string, bool> ShowCoresMenu(List<Core> cores, string message, bool isCoreSelection)
-    {
-        const int pageSize = 12;
-        var offset = 0;
-        bool more = true;
-        var results = new Dictionary<string, bool>();
-        string save,
-            quit;
-
-        if (isCoreSelection)
-        {
-            save = "Save Choices";
-            quit = "Quit without saving";
-        }
-        else
-        {
-            save = "Apply Choices";
-            quit = "Quit without applying";
-        }
-
-        while (more)
-        {
-            var menu = new ConsoleMenu()
-                .Configure(config =>
-                {
-                    config.Selector = "=>";
-                    config.EnableWriteTitle = false;
-                    config.WriteHeaderAction = () => Console.WriteLine($"{message} Use enter to check/uncheck your choices.");
-                    config.SelectedItemBackgroundColor = Console.ForegroundColor;
-                    config.SelectedItemForegroundColor = Console.BackgroundColor;
-                    config.WriteItemAction = item => Console.Write("{0}", item.Name);
-                });
-            var current = -1;
-
-            if ((offset + pageSize) <= cores.Count)
-            {
-                menu.Add("Next Page", thisMenu =>
-                {
-                    offset += pageSize;
-                    thisMenu.CloseMenu();
-                });
-            }
-
-            foreach (Core core in cores)
-            {
-                current++;
-
-                if ((current <= (offset + pageSize)) && (current >= offset))
-                {
-                    var coreSettings = ServiceHelper.SettingsService.GetCoreSettings(core.identifier);
-                    var selected =
-                        (isCoreSelection && !coreSettings.skip) ||
-                        (results.TryGetValue(core.identifier, out var result) && result);
-                    var name = core.identifier;
-                    var title = MenuItemName(name, selected, core.requires_license);
-
-                    menu.Add(title, thisMenu =>
-                    {
-                        selected = !selected;
-
-                        if (results.ContainsKey(core.identifier))
-                        {
-                            results[core.identifier] = selected;
-                        }
-                        else
-                        {
-                            results.Add(core.identifier, selected);
-                        }
-
-                        thisMenu.CurrentItem.Name = MenuItemName(core.identifier, selected, core.requires_license);
-                    });
-                }
-            }
-
-            if ((offset + pageSize) <= cores.Count)
-            {
-                menu.Add("Next Page", thisMenu =>
-                {
-                    offset += pageSize;
-                    thisMenu.CloseMenu();
-                });
-            }
-
-            if (offset != 0)
-            {
-                menu.Add("Prev Page", thisMenu =>
-                {
-                    offset -= pageSize;
-                    thisMenu.CloseMenu();
-                });
-            }
-
-            menu.Add(save, thisMenu =>
-            {
-                thisMenu.CloseMenu();
-                more = false;
-            });
-
-            menu.Add(quit, thisMenu =>
-            {
-                thisMenu.CloseMenu();
-                results.Clear();
-                more = false;
-            });
-
-            menu.Show();
-        }
-
-        return results;
-    }
-
-    private static Dictionary<string, bool> RunCoreSelector(List<Core> cores, string message = "Select your cores.")
-    {
-        Dictionary<string, bool> results = null;
-
-        if (ServiceHelper.SettingsService.GetConfig().download_new_cores?.ToLowerInvariant() == "yes")
-        {
-            foreach (Core core in cores)
-            {
-                ServiceHelper.SettingsService.EnableCore(core.identifier);
-            }
-        }
-        else
-        {
-            results = ShowCoresMenu(cores, message, true);
-
-            foreach (var item in results)
-            {
-                if (item.Value)
-                    ServiceHelper.SettingsService.EnableCore(item.Key);
-                else
-                    ServiceHelper.SettingsService.DisableCore(item.Key);
-            }
-        }
-
-        ServiceHelper.SettingsService.Save();
-
-        return results;
-    }
-
-    private static void PlatformImagePackSelector()
-    {
-        Console.Clear();
-
-        if (ServiceHelper.PlatformImagePacksService.List.Count > 0)
-        {
-            int choice = 0;
-            var menu = new ConsoleMenu()
-                .Configure(config =>
-                {
-                    config.Selector = "=>";
-                    config.EnableWriteTitle = false;
-                    config.WriteHeaderAction = () => Console.WriteLine("So, what'll it be?:");
-                    config.SelectedItemBackgroundColor = Console.ForegroundColor;
-                    config.SelectedItemForegroundColor = Console.BackgroundColor;
-                });
-
-            foreach (var pack in ServiceHelper.PlatformImagePacksService.List)
-            {
-                menu.Add($"{pack.owner}: {pack.repository} {pack.variant ?? string.Empty}", thisMenu =>
-                {
-                    choice = thisMenu.CurrentItem.Index;
-                    thisMenu.CloseMenu();
-                });
-            }
-
-            menu.Add("Go Back", thisMenu =>
-            {
-                choice = ServiceHelper.PlatformImagePacksService.List.Count;
-                thisMenu.CloseMenu();
-            });
-
-            menu.Show();
-
-            if (choice < ServiceHelper.PlatformImagePacksService.List.Count && choice >= 0)
-            {
-                ServiceHelper.PlatformImagePacksService.Install(
-                    ServiceHelper.PlatformImagePacksService.List[choice].owner,
-                    ServiceHelper.PlatformImagePacksService.List[choice].repository,
-                    ServiceHelper.PlatformImagePacksService.List[choice].variant);
-            }
-            else if (choice == ServiceHelper.PlatformImagePacksService.List.Count)
-            {
-                // What causes this?
-            }
-            else
-            {
-                Console.WriteLine("You fucked up.");
-            }
-        }
-        else
-        {
-            Console.WriteLine("None found. Have a nice day.");
-        }
-    }
-
-    private static void DisplayModeSelector(bool showCoreSelector = false)
-    {
-        Console.Clear();
-
-        const int pageSize = 12;
-        var offset = 0;
-        var more = true;
-        var count = 0;
-        var results = new List<string>();
-        var allDisplayModes = ServiceHelper.CoresService.GetAllDisplayModes();
-
-        while (more)
-        {
-            var menu = new ConsoleMenu()
-                .Configure(config =>
-                {
-                    config.Selector = "=>";
-                    config.EnableWriteTitle = false;
-                    config.WriteHeaderAction = () =>
-                    {
-                        Console.WriteLine("Which display modes would you like to enable?");
-                        Console.WriteLine($"Note: There is a maximum of 16. You have {16 - count} remaining.");
-                    };
-                    config.SelectedItemBackgroundColor = Console.ForegroundColor;
-                    config.SelectedItemForegroundColor = Console.BackgroundColor;
-                    config.WriteItemAction = item => Console.Write("{0}", item.Name);
-                });
-            var current = -1;
-
-            if ((offset + pageSize) <= allDisplayModes.Count)
-            {
-                menu.Add("Next Page", thisMenu =>
-                {
-                    offset += pageSize;
-                    thisMenu.CloseMenu();
-                });
-            }
-
-            foreach (DisplayMode displayMode in allDisplayModes)
-            {
-                current++;
-
-                if ((current <= (offset + pageSize)) && (current >= offset))
-                {
-                    var selected = results.Contains(displayMode.value);
-                    var title = MenuItemName(displayMode.description, selected);
-
-                    menu.Add(title, thisMenu =>
-                    {
-                        if (count >= 16 && !selected)
-                            return;
-
-                        selected = !selected;
-
-                        if (selected)
-                        {
-                            results.Add(displayMode.value);
-                            count++;
-                        }
-                        else
-                        {
-                            results.Remove(displayMode.value);
-                            count--;
-                        }
-
-                        thisMenu.CurrentItem.Name = MenuItemName(displayMode.description, selected);
-                    });
-                }
-            }
-
-            if ((offset + pageSize) <= allDisplayModes.Count)
-            {
-                menu.Add("Next Page", thisMenu =>
-                {
-                    offset += pageSize;
-                    thisMenu.CloseMenu();
-                });
-            }
-
-            if (offset != 0)
-            {
-                menu.Add("Prev Page", thisMenu =>
-                {
-                    offset -= pageSize;
-                    thisMenu.CloseMenu();
-                });
-            }
-
-            if (!showCoreSelector)
-            {
-                menu.Add("Apply Choices", thisMenu =>
-                {
-                    EnableDisplayModes(displayModes: results.ToArray());
-                    thisMenu.CloseMenu();
-                    more = false;
-                });
-            }
-            else
-            {
-                menu.Add("Select Cores", thisMenu =>
-                {
-                    thisMenu.CloseMenu();
-                    more = false;
-
-                    var coreResults = ShowCoresMenu(
-                        ServiceHelper.CoresService.InstalledCores,
-                        "Which cores would you like to apply the selected display modes to?",
-                        false);
-
-                    EnableDisplayModes(
-                        coreResults.Where(kvp => kvp.Value).Select(kvp => kvp.Key).ToList(),
-                        results.ToArray());
-                });
-            }
-
-            menu.Add("Quit without applying", thisMenu =>
-            {
-                thisMenu.CloseMenu();
-                more = false;
-            });
-
-            menu.Show();
-        }
-    }
-
-    private static void SettingsMenu()
-    {
-        Console.Clear();
-
-        var type = typeof(Config);
-        var menuItems =
-            from property in type.GetProperties()
-            let attribute = property.GetCustomAttributes(typeof(DescriptionAttribute), true)
-            where attribute.Length == 1
-            select (property.Name, ((DescriptionAttribute)attribute[0]).Description);
-        var menu = new ConsoleMenu()
-            .Configure(config =>
-            {
-                config.Selector = "=>";
-                config.EnableWriteTitle = false;
-                config.WriteHeaderAction = () => Console.WriteLine("Settings. Use enter to check/uncheck your choices.");
-                config.SelectedItemBackgroundColor = Console.ForegroundColor;
-                config.SelectedItemForegroundColor = Console.BackgroundColor;
-                config.WriteItemAction = item => Console.Write("{0}", item.Name);
-            });
-
-        foreach (var (name, text) in menuItems)
-        {
-            var property = type.GetProperty(name);
-            var value = (bool)property!.GetValue(ServiceHelper.SettingsService.GetConfig())!;
-            var title = MenuItemName(text, value);
-
-            menu.Add(title, thisMenu =>
-            {
-                value = !value;
-                property.SetValue(ServiceHelper.SettingsService.GetConfig(), value);
-                thisMenu.CurrentItem.Name = MenuItemName(text, value);
-            });
-        }
-
-        menu.Add("Save", thisMenu => { thisMenu.CloseMenu(); });
-
-        menu.Show();
-
-        ServiceHelper.SettingsService.Save();
     }
 
     private static string MenuItemName(string title, bool value, bool requiresLicense = false)
