+++ src/services/CoresService.Helpers.cs
@@ -63,7 +63,7 @@ public partial class CoresService
         ZipHelper.ExtractToDirectory(zipPath, tempDir, true);
 
         // Clean problematic directories and files.
-        Util.CleanDir(tempDir, this.installPath, this.settingsService.GetConfig().preserve_platforms_folder, false, platformId);
+        Util.CleanDir(tempDir, this.installPath, this.settingsService.GetConfig().preserve_platforms_folder, platformId);
 
         // Move the files into place and delete our core's temp directory.
         WriteMessage("Installing...");
